Run all styling, documentation, and testing for FIMS repository

FIMS consists of both C++ and R code but no function exists to style, document, and test both file types. This function first works with the C++ code and then moves onto the R code. For each file type it styles the files, e.g., puts in appropriate spacing between brackets; then it build the documentation, e.g., roxygen2::roxygenize(); and then it tests the functions.

Usage

check_fims()

Value

An invisible list is returned with the following four items:

style_r: Contains the report from styler::style_pkg().

spelling_code: Contains a 2-column data frame of words that are spelled incorrectly in the code but not in the excluded word list.

spelling_package: Contains a 2-column data frame of words that need to be added to the WORDLIST file. This will be an empty data frame if no words are spelled incorrectly.

spelling_word_list: A vector of strings indicating the words that are included in the WORDLIST file and are ignored when performing a spell check.

Examples

if (FALSE) { # \dontrun{
check_results <- check_fims()
} # }