Setup and run the google test suite

Intended for developers to set up their local environment and run the google test suite from R.

Usage

setup_and_run_gtest(...)

Arguments

...

Additional arguments to ctest --test-dir build such as "--rerun-failed --output-on-failure".