Class constructors for FIMSFrame and associated child classes

All constructor functions take a single input and build an object specific to the needs of each model type within FIMS. FIMSFrame is the parent class and the associated child classes have additional slots needed for each model type.

Usage

FIMSFrame(data)

Arguments

data

A data.frame that contains the necessary columns to construct a data frame of a given FIMSFrame-class.

Value

An object of the S4 class FIMSFrame or one of its child classes is validated and then returned. All objects will at a minimum have a slot called data to store the input data frame. Additional slots are dependent on the child class. Use showClass() to see all available slots.